fix(prediction): DIA getValue uint128 return type + 18-dec regression guard
Empirically confirmed against the live LitVM DIA feed
(0x49c39225Dbc64700936bb641d1E81113DbadD2DF): getValue("BTC/USD") returned
62865495019032875630592 (~$62,865 x 1e18), proving 18 decimals — DIA's docs
claiming 8 are wrong for this deployment.

1. IDIAOracle.getValue return type corrected uint256 -> uint128 to match the
   real ABI. Adapter is unchanged structurally: the uint128 value widens to
   uint256 on destructuring and the over-int256.max clamp stays as pure
   defense-in-depth (uint128 max ~3.4e38 << int256 max, so it can no longer
   fire). MockDIAOracle mirrors the uint128 return; stored values still fit.

2. Added test_LatestRoundData_RealisticBtcPrice_NoScalingError: pins a
   realistic $62,865 (62_865e18) price mapping through the adapter exactly,
   provably off by no power of ten, both directly and through
   SafeAggregatorReader. decimals()==18 already correct — now regression-locked.

Second getValue return remains the freshness TIMESTAMP (-> updatedAt/startedAt
staleness anchor), never a price component. No perp money-path files touched.
forge test: 403 passed, 0 failed.

+++ src/prediction/IDIAOracle.sol
@@ -6,7 +6,7 @@ pragma solidity ^0.8.24;
  * @notice The native surface of DIA's custom LitVM oracle (live on testnet at
  *         0x49c39225Dbc64700936bb641d1E81113DbadD2DF). It is NOT AggregatorV3-shaped:
  *         it keys prices by a `"SYMBOL/USD"` string and returns a plain
- *         `(uint256 value, uint256 timestamp)` tuple. {DIAAggregatorV3Adapter}
+ *         `(uint128 value, uint128 timestamp)` tuple. {DIAAggregatorV3Adapter}
  *         wraps one of these keys per asset and re-exposes it as {IAggregatorV3},
  *         so the prediction stack keeps reading through a single oracle abstraction
  *         and the proven market contracts stay untouched.
@@ -17,8 +17,11 @@ interface IDIAOracle {
     /**
      * @notice Latest price for a `"SYMBOL/USD"` key (e.g. `"BTC/USD"`).
      * @param key        The asset pair, formatted `"SYMBOL/USD"`.
-     * @return value     USD price scaled by 18 decimals. `0` signals no/void value.
+     * @return value     USD price scaled by 18 decimals (confirmed empirically against the
+     *                   live LitVM feed — a BTC read returned 62865495019032875630592, i.e.
+     *                   ~$62,865 x 1e18). `0` signals no/void value. Packed as `uint128`
+     *                   on-chain: 18-dec USD prices never approach `type(uint128).max`.
      * @return timestamp Unix seconds (UTC) of the last update — the staleness anchor.
      */
-    function getValue(string memory key) external view returns (uint256 value, uint256 timestamp);
+    function getValue(string memory key) external view returns (uint128 value, uint128 timestamp);
 }
